+++ src/test/java/codesquad/fineants/spring/api/portfolio_notification/MailServiceTest.java
@@ -10,6 +10,7 @@ import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 
 import codesquad.fineants.spring.AbstractContainerBaseTest;
+import codesquad.fineants.spring.api.mail.MailService;
 
 class MailServiceTest extends AbstractContainerBaseTest {
